+++ pages/_app.tsx
@@ -3,7 +3,7 @@ import { createTheme, ThemeProvider } from '@mui/material';
 import CssBaseline from '@mui/material/CssBaseline';
 import type { AppProps } from 'next/app';
 import * as React from 'react';
-import Navbar from './components/navbar';
+import Navbar from './components/NavBar';
 
 import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
@@ -11,7 +11,7 @@ import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
 
 import '../styles/globals.css';
-import theme from '../styles/theme/theme';
+import theme from '../styles/theme/Theme';
 import createEmotionCache from '../utility/createEmotionCache';
 interface MyAppProps extends AppProps {
   emotionCache?: EmotionCache;
