+++ pages/index.tsx
@@ -12,7 +12,7 @@ const Home: NextPage = () => {
     <div>
       <Head>
         <title>Complanion</title>
-        <meta name="description" content="Trip planning application enerated by create next app" />
+        <meta name="description" content="Trip planning application generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
